requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask_login
flask_migrate
python-dotenv
